refactor(platform): shared pagination, sort allow-lists, SQLSTATE mapping, LIKE escaping (#47)
* refactor(platform): shared pagination, sort allow-lists, SQLSTATE mapping, LIKE escaping

Phase 2 of docs/remediation-plan.md. Adds four small platform modules and
migrates every domain module onto them:

- `platform::pagination`: `clamp` / `clamp_with` (page >= 1, page_size
  default 20 and capped at 100; audit keeps 50/200), saturating `offset`,
  and one `Page<T>` envelope. Replaces 18 `clamp_page` copies, 19 `Page<T>`
  structs and 38 `(page - 1) * page_size` sites. Fixes the uncapped
  `page_size` in labels and reporting and the overflow panic on
  `?page=i64::MAX`. `library` and `water` move from i32 to i64 paging.
- `platform::sort`: `parse` / `parse_with(Nulls)` over per-module
  `const *_ALLOWED_SORT` tables. Replaces all 27 hand-written resolvers.
  Unknown sort fields now return 400 everywhere (14 resolvers used to fall
  back silently). Repository list functions that parse a sort now return
  `ApiError`.
- `platform::errors`: single `is_unique_violation` (was defined 11 times)
  and SQLSTATE mapping in `From<sqlx::Error>`: 23505 -> 409, 23514/23503/
  22001/22P02 -> 400 validation instead of 500.
- `platform::sql`: `like_contains` / `like_prefix` escape `\ % _`
  (Postgres's default escape char is backslash, so no ESCAPE clause). Used
  at all 15 LIKE/ILIKE sites.
- `validation_to_api_error` is now `pub(crate)` for Phase 6.

Tests: 31 unit tests for the new modules; integration tests for page_size
clamping, i64::MAX page, negative page, unknown sort on two endpoints, and
a CHECK violation returning 400.


* docs(plan): tick Phase 2


* refactor(platform): clippy fallout from ApiError return types

needless_question_mark, useless_conversion, and redundant format! refs
introduced by the sort/pagination migration.


* refactor(platform): clippy fallout from ApiError return types

needless_question_mark, useless_conversion, and redundant format! refs
introduced by the sort/pagination migration.


---------

+++ src/water/handler.rs
@@ -57,8 +57,8 @@ pub fn routes(state: AppState) -> Router {
 #[derive(Debug, Deserialize)]
 struct ProfileQuery {
     sort: Option<String>,
-    page: Option<i32>,
-    page_size: Option<i32>,
+    page: Option<i64>,
+    page_size: Option<i64>,
 }
 
 #[derive(Debug, Deserialize)]
@@ -66,8 +66,8 @@ struct AdjustmentQuery {
     batch_id: Option<Uuid>,
     recipe_id: Option<Uuid>,
     sort: Option<String>,
-    page: Option<i32>,
-    page_size: Option<i32>,
+    page: Option<i64>,
+    page_size: Option<i64>,
 }
 
 // ---- Profiles ----
